Enhance Niagara renderer with scene management and shader updates
- Added a dropdown menu for scene selection in the GUI, allowing users to switch between predefined test scenes.
- Updated the scene loading mechanism to utilize a vector of scene paths, improving flexibility in scene management.
- Introduced a maximum texture limit to optimize texture handling during rendering.
- Refactored texture upload logic to accommodate the new maximum texture constraint.
- Adjusted shader structures to support updated data types for meshlet properties, enhancing compatibility with the rendering pipeline.
- Improved camera property handling during scene conversion, ensuring accurate scene representation.

+++ Source/Niagara/NiagaraScene.h
@@ -13,15 +13,15 @@ struct alignas(8) NiagaraMeshlet
 {
     uint16_t center[3];
     uint16_t radius;
-    int8_t cone_axis[3];
-    int8_t cone_cutoff;
+    int16_t cone_axis[3];
+    int16_t cone_cutoff;
 
     uint32_t dataOffset; // dataOffset..dataOffset+vertexCount-1 stores vertex indices, we store indices packed in 4b units after that
     uint32_t baseVertex;
-    uint8_t vertexCount;
-    uint8_t triangleCount;
-    uint8_t shortRefs;
-    uint8_t padding;
+    uint16_t vertexCount;
+    uint16_t triangleCount;
+    uint16_t shortRefs;
+    uint16_t padding;
 };
 
 struct alignas(16) NiagaraMaterial
